moving methods for reuse
