+++ tests/unit/AbstractLiveMailboxTest.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 
 namespace PhpImap;
 
+use Generator;
 use ParagonIE\HiddenString\HiddenString;
 use PHPUnit\Framework\TestCase;
 
@@ -21,6 +22,17 @@ use PHPUnit\Framework\TestCase;
  *	3:string,
  *	4?:string
  * }
+ * @psalm-type COMPOSE_ENVELOPE = array{
+ *	subject?:string
+ * }
+ * @psalm-type COMPOSE_BODY = list<array{
+ *	type?:int,
+ *	encoding?:int,
+ *	charset?:string,
+ *	subtype?:string,
+ *	description?:string,
+ *	disposition?:array{filename:string}
+ * }>
  */
 abstract class AbstractLiveMailboxTest extends TestCase
 {
@@ -42,6 +54,118 @@ abstract class AbstractLiveMailboxTest extends TestCase
         }
 
         return $sets;
+    }
+
+    /**
+     * @psalm-return Generator<int, array{0:COMPOSE_ENVELOPE, 1:COMPOSE_BODY, 2:string}, mixed, void>
+     */
+    public function ComposeProvider(): Generator
+    {
+        yield from [];
+    }
+
+    /**
+     * @psalm-return Generator<int, array{
+     *	0:MAILBOX_ARGS,
+     *	1:COMPOSE_ENVELOPE,
+     *	2:COMPOSE_BODY,
+     *	3:string,
+     *	4:bool
+     * }, mixed, void>
+     */
+    public function AppendProvider(): Generator
+    {
+        foreach ($this->MailBoxProvider() as $mailbox_args) {
+            foreach ($this->ComposeProvider() as $compose_args) {
+                [$envelope, $body, $expected_compose_result] = $compose_args;
+
+                yield [$mailbox_args, $envelope, $body, $expected_compose_result, false];
+            }
+
+            foreach ($this->ComposeProvider() as $compose_args) {
+                [$envelope, $body, $expected_compose_result] = $compose_args;
+
+                yield [$mailbox_args, $envelope, $body, $expected_compose_result, true];
+            }
+        }
+    }
+
+    /**
+     * @dataProvider AppendProvider
+     *
+     * @group live
+     *
+     * @depends testGetImapStream
+     * @depends testMailCompose
+     *
+     * @psalm-param MAILBOX_ARGS $mailbox_args
+     * @psalm-param COMPOSE_ENVELOPE $envelope
+     * @psalm-param COMPOSE_BODY $body
+     */
+    public function testAppend(
+        array $mailbox_args,
+        array $envelope,
+        array $body,
+        string $_expected_compose_result,
+        bool $pre_compose
+    ): void {
+        if ($this->MaybeSkipAppendTest($envelope)) {
+            return;
+        }
+
+        list($search_criteria) = $this->SubjectSearchCriteriaAndSubject($envelope);
+
+        list($mailbox, $remove_mailbox, $path) = $this->getMailboxFromArgs(
+            $mailbox_args
+        );
+
+        $search = $mailbox->searchMailbox($search_criteria);
+
+        $this->assertCount(
+            0,
+            $search,
+            (
+                'If a subject was found,'.
+                ' then the message is insufficiently unique to assert that'.
+                ' a newly-appended message was actually created.'
+            )
+        );
+
+        $message = [$envelope, $body];
+
+        if ($pre_compose) {
+            $message = Imap::mail_compose($envelope, $body);
+        }
+
+        $mailbox->appendMessageToMailbox($message);
+
+        $search = $mailbox->searchMailbox($search_criteria);
+
+        $this->assertCount(
+            1,
+            $search,
+            (
+                'If a subject was not found, '.
+                ' then Mailbox::appendMessageToMailbox() failed'.
+                ' despite not throwing an exception.'
+            )
+        );
+
+        $mailbox->deleteMail($search[0]);
+
+        $mailbox->expungeDeletedMails();
+
+        $mailbox->switchMailbox($path->getString());
+        $mailbox->deleteMailbox($remove_mailbox);
+
+        $this->assertCount(
+            0,
+            $mailbox->searchMailbox($search_criteria),
+            (
+                'If a subject was found,'.
+                ' then the message is was not expunged as requested.'
+            )
+        );
     }
 
     /**
@@ -105,5 +229,18 @@ abstract class AbstractLiveMailboxTest extends TestCase
 
         /** @psalm-var array{0:string, 1:string} */
         return [$search_criteria, (string) $subject];
+    }
+
+    protected function MaybeSkipAppendTest(array $envelope): bool
+    {
+        if (!isset($envelope['subject'])) {
+            $this->markTestSkipped(
+                'Cannot search for message by subject, no subject specified!'
+            );
+
+            return true;
+        }
+
+        return false;
     }
 }
